requirements.txt
# django
django==3.0.8
django-environ==0.4.5
mysqlclient==1.4.6
numpy==1.15.4
pillow==7.1.2
django-crispy-forms==1.9.1

flake8
pep8-naming